Sending first data now hacky af
But it works. UI is not updating anything
at this point, but we *are* sending
and receiving between gdb and our
own little DA. It's hacky right now.
It's bad right now. It lacks 99.9%
of all features.

This is just to learn what to do
and where to do it.