Threaded cpp project now does multi threaded things
Basically a simple mandelbrot calc over a huge
"screen". The reason why it's so large, is because
sometimes we want to be able to hit a debug breakpoint
inside the extension and *not* have the thread
finish it's executition before we've decided (in real time)
what to do and what to read.