VScode now knows about threads
Some features are yet to be implemented though.

For instance; evaluateRequest isn't implemented at all yet,
but when it is, we will need to track for what thread,
the "hovered over" variable is supposed to be evaluated for.

Also cleaned up the display of the threads. The number
between () is the Light-weight process ID.