Correctly handles non-stop mode now.
When a thread hits a breakpoint, it now is correctly shown as
"paused on breakpoint" in the call stack, while other live threads
are shown as "paused" (since they can be paused at arbitrary locations).

This solves the issue of VSCode not knowing *what* thread hit a breakpoint.
This was caused by gdb running in non-stop, which means when
a thread hits a breakpoint, it does *not* tell GDB to stop the other ones.