VSCode requesting stack frames is a rats' nest. For some inexplicable reason, it sends stackFrame requests twice per step loop.
For us, it isn't all that bad, we we read the stack pointer and the frame address,
verify that we remained in the same stack frame as before and just return previous result;
if we didn't we build the stack frame stack and return it. So for the general use case
i.e. stepping lines, the cost is held low.