External Console management + formatting
Since new terminals use the client/server architecture
we have to force whatever terminal is being used to spawn a new process
This requires knowledge of the individual terminals and will be needed to be added
as we go along.

For now, we know gnome-terminal and tilix, terminals like xterm do not operate
like that and thus have no problems here.

We resolve the symlink if it is one, and then check basename of the file
and lookup the parameter required to spawn a new process.