Refactor of the spawning of the gdb process
Adds an abstract SpawnConfig base class which the
different configurations inherit from.

The base class main job is to produce 2 parameters in their
finalaized form, the path to GDB and all the potential parameters.

These vary across debug session requests, but also if it's an rr session
or not.

Much cleaner implementation