Fixed bug due to Python API returning wrong data
Getting the symbol table from a function, using the following
fnSymbolTable = gdb.newest_frame().function().symtab

returns the wrong data if the debuggee is an optimized build and
possibly where inlining is concerned, although it's not entirely obvious
that this is the case. It might just be buggy.

Instead, retrieving the symboltable directly from the
`frame.find_sal()` function (and accessing the member symtab) gets
the correct result. Investigation is needed into the Python API of gdb
to find out whether this is a bug or a feature.