Fixed bug introduced by fix of #133
Problem with MI "protocol" pop up, yet again.

MI has a frustrating property that the messages it sends isn't well
defined, and as such, sometimes an onExec might emit certain properties
and fields, while other times it does not. This forces client code to
always check if field F exists.

This time it was the "reason" field.