Updated APT script
It now waits for response from Midas on what packages it should install, instead of hard coding.

This requires some additional logic but is certainly worth it. N.B. this script still needs sudo to run,
so will not solve the problem of only _checking_ for missing dependencies
so that has to be tackled separately.