Make launch look for Midas environment variables (the "cache")
Unless user has explicitly specified an RR path, we look in Midas' own settings. If the user has acquired RR by building it using Midas, we will know the exact path to use and use that.

The default value of the cache/"midas environment variable " is "rr" thus, it would default to looking into $PATH.