Added rr path to User/Workspace settings
This will override _all_ other RR paths including launch.json and the cached path, set by the build-rr functionality

Added some error handling and cleaned up trace workspace picking.

We don't want to display "cpu_lock" for instance. And if the `rr ls` command only returns "cpu_lock" and "latest-trace" it means the RR dir is empty which we will stop the session from starting and reporting this to the user.