added whereis command
Since vscode might be part of an IDE-like environment, the user might be in
a VIRTUAL_ENV set by Python for instance.

In these case `which` will return the path of that binary, which will be no good for us.

whereis OTOH returns all items in $PATH and returns them as a string array.

run_install now calls whereis instead if it can find VIRTUAL_ENV amongst the environment variables.