Refactored config cache
The previous version using VSCode mementos was a terrible system. The interface with the key-value pair database that VSCode provides, provides no means (at all) at versioning, or maintenance which makes that data impossible for us to verify or keep sane.

Reasons for that is:
- What version is the key-value pair?
- It doesn't get removed or cleaned up if extension gets uninstalled
- No way to sanely check if value has been set (yes, technically a default value can be provided but what if a value should be set to undefined?)

These are some of the reasons but I came across others.