Refactor
Use Midas API resolve of tools.
Use new exceptions to be thrown during failure of resolveConfig.

With the new "exceptions-like" objects being thrown it's easier to follow what's going on.

It's also easier to change the types being throw, going forward (for instance, now they're strings, in the future we'll just make them integer constants to swtich on err.type)