- Added output channel loggers to the API - Setup logging using MidasAPI-logger API - startWithRR can launch remote RR sessions - added Remote RR session & config in RemoteRRSpawnConfig
Removed serverAddress from launch.json config,
as it's replaced by remoteTargetConfig now - and this field
signals whether or not a session is a remote or local session.

This field exists on all session types, and works the same way;
if it's defined, it will be a remote session on a remote target.