Added find_value to VariablesReference
Use in the future, for databreakpointInfo requests. This way we can search "UI containers" for a specific name and resolve the internal GDB value for that name.

Databreakpoints are set using "dataId" in DAP, and for our case, our unique id will have to be the address of the value. It's not exactly human friendly when it comes to debugging, but it will be unique. It's a possibility it will be called something horrible in the UI, but I'm not sure.