Fixes #188
Midas couldn't be started because the list command used was `rr ls -l`
which essentially does a `rr ps` per each record in trace directory, and
since a record is in-flight, the command fails entirely. This is
probably more of an "error bug" (or ux flaw) but, we manage it by just
saying `rr ls` instead.