Spring cleaning
Moved all constants I could find into constants.js and removed any other files that just consistent of constants. Removed contents of launch.json. Users should add this to a local ignore from here on.

Changed the name of the midas mode that runs against the native debugger/backend, to "midas-native".

Made the resolveDebugConfiguration just return the config directly and we do validation on the one with the variables substituted. Why this isn't/wasn't default documented behavior, i will never know. But that means now, ${workspaceFolder} can be used everywhere.

Ran formatting on the files changed, as this had previously been missed. But much to due to how javascript tooling is a shit show - who knows if I actually ran it with the correct settings.