New commands
Added continueAll
Added pauseAll
Deprecated allStopMode as config item, it's now called
noSingleThreadControl.

Clicking frames in the call stack will also make sure
the gdb backend has "selected thread" as this
can be confusing when entering commands in the debug console
that it's not operating on the thread the user think it is.