Removed old implementation entirely
This commit is a rather large "trust me bro", but it had to be made.

It removes support for the old midas implementation before we had our
own DAP implementation, and we used GDB/MI + gdb-js and converted that
to a DAP-format for communcation.

Additionally, I've added `prettier` as a dev-dependency, and as a script
command `npm run format` - so that potential contributors don't have to
deal with the headache of previous formatting. It also had to be done
now, so ripped the band aid off.